Move clinic profile editing to dedicated route
