+++ app/components/navbar/navbar.js
@@ -144,6 +144,7 @@ export default translate()(class extends React.Component {
           '/clinic-admin',
           '/prescriptions',
           '/clinic-invite',
+          '/clinic-profile',
           '/clinic-workspace',
           '/clinic-workspace/patients',
           '/clinic-workspace/invites',
