+++ app/pages/clinicadmin copy/clinicadmin.js
